Move JuColumnTransformer and make setting default of apply to in PipelienCreator possible
